+++ metadata.rb
@@ -4,7 +4,7 @@ issues_url       'https://github.com/sous-chefs/atlantis/issues'
 maintainer       'Sous Chefs'
 maintainer_email 'help@sous-chefs.org'
 chef_version     '>= 15.3'
-license          'MIT'
+license          'Apache-2.0'
 description      'Installs and configures atlantis'
 version          '2.0.4'