+++ test/cookbooks/test/metadata.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-name             'test'
-aintainer        'Sous Chefs'
+name 'test'
+maintainer 'Sous Chefs'
 maintainer_email 'help@sous-chefs.org'
-license 'MIT'
+license           'Apache-2.0'
 description       'Testing cookbook for atlantis'
 version           '0.0.1'